feat: memberId's data type replace Integer with Long

+++ src/main/java/com/marykuo/security/adapter/in/api/controller/member/query/QuerySingleMemberController.java
@@ -28,7 +28,7 @@ public class QuerySingleMemberController {
     @GetMapping(value = "/v1" + MEMBER_SINGLE)
     public ResponseEntity<DataResponse<QuerySingleMemberResponse>> query(
             @RequestAttribute(LOGIN_MEMBER) MemberEntity loginMember,
-            @PathVariable("memberId") Integer memberId
+            @PathVariable("memberId") Long memberId
     ) {
         log.debug("member[{}] query member[{}]", loginMember.getMemberId(), memberId);
